pyproject.toml
[project]
name = "mcp-google-docs"
version = "0.2.13"
description = "mcp server for viewing, editing, creating, google docs (supports comments)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.160.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "httplib2>=0.22.0",
    "mcp>=1.2.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.4",
]

[tool]
uv.package = true

[project.scripts]
server = "mcp_server.mcp_server:main"
